Added a namespace for the vk wrapper
