+++ include/Raytracer.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include "VoxelWorld.hpp"
-#include "image.hpp"
+#include "VkZero/image.hpp"
 #include "shaders.hpp"
 #include <memory>
 #include <vector>
@@ -12,54 +12,58 @@
 #define RAYTRACE_WIDTH 1920
 #define RAYTRACE_HEIGHT 1080
 
+namespace VkZero
+{
+  class VulkanContext;
+}
 class Raytracer {
 public:
   Raytracer(std::unique_ptr<class CommandManager> &commandManager,
-            std::unique_ptr<class VulkanContext> &vulkanContext,
+            std::unique_ptr<VkZero::VulkanContext> &vulkanContext,
             std::unique_ptr<VoxelWorld> &voxelWorld,
             std::unique_ptr<class Camera> &camera);
   ~Raytracer();
 
   void createRaytracingPipeline(
       VkDevice device, std::vector<VkBuffer> &uniformBuffer,
-      StagedSharedImage* voxelImage,
-      StagedSharedImage* voxelChunkMapImage);
+      VkZero::StagedSharedImage* voxelImage,
+      VkZero::StagedSharedImage* voxelChunkMapImage);
   void createRaytracingResources(
       std::unique_ptr<class CommandManager> &commandManager,
-      std::unique_ptr<class VulkanContext> &vulkanContext);
+      std::unique_ptr<VkZero::VulkanContext> &vulkanContext);
   void recordRaytracingCommandBuffer(VkCommandBuffer commandBuffer,
                                      uint32_t imageIndex, uint8_t section);
 
   const VkPipeline &getPipeline() const;
   const VkPipelineLayout &getPipelineLayout() const;
   const VkDescriptorSet &getDescriptorSet(int i) const;
-  SwapImage* getStorageImage();
+  VkZero::SwapImage* getStorageImage();
 
 // private:
 
-  using RaytracingPushConstants = ShaderPushConstants<PushConstant<RaytracingPushConstant, SHADER_RGEN>>;
-  using RaytracingShaderGroup = ShaderGroup<RaytracingPushConstants, main_rmiss, main_rgen>;
-  using RaytracingResourceSet = ShaderResourceSet<
-    ResourceBinding<SwapImage, VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, SHADER_RGEN, 0, 1>,
-    ResourceBinding<SwapImage, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, SHADER_RGEN, 1, 1>,
-    ResourceBinding<SwapImage, VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, SHADER_RGEN, 2, 1>,
-    ResourceBinding<StagedSharedImage, VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, SHADER_RGEN, 3, 512>,
-    ResourceBinding<StagedSharedImage, VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, SHADER_RGEN, 4, 1>,
-    ResourceBinding<SwapImage, VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, SHADER_RGEN, 5, 1>,
-    ResourceBinding<SwapImage, VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, SHADER_RGEN, 6, 1>,
-    ResourceBinding<SwapImage, VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, SHADER_RGEN, 7, 1>,
-    ResourceBinding<SwapImage, VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, SHADER_RGEN, 8, 1>
+  using RaytracingPushConstants = VkZero::ShaderPushConstants<VkZero::PushConstant<RaytracingPushConstant, VkZero::SHADER_RGEN>>;
+  using RaytracingShaderGroup = VkZero::ShaderGroup<RaytracingPushConstants, main_rmiss, main_rgen>;
+  using RaytracingResourceSet = VkZero::ShaderResourceSet<
+    VkZero::ResourceBinding<VkZero::SwapImage, VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, VkZero::SHADER_RGEN, 0, 1>,
+    VkZero::ResourceBinding<VkZero::SwapImage, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, VkZero::SHADER_RGEN, 1, 1>,
+    VkZero::ResourceBinding<VkZero::SwapImage, VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, VkZero::SHADER_RGEN, 2, 1>,
+    VkZero::ResourceBinding<VkZero::StagedSharedImage, VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, VkZero::SHADER_RGEN, 3, 512>,
+    VkZero::ResourceBinding<VkZero::StagedSharedImage, VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, VkZero::SHADER_RGEN, 4, 1>,
+    VkZero::ResourceBinding<VkZero::SwapImage, VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, VkZero::SHADER_RGEN, 5, 1>,
+    VkZero::ResourceBinding<VkZero::SwapImage, VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, VkZero::SHADER_RGEN, 6, 1>,
+    VkZero::ResourceBinding<VkZero::SwapImage, VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, VkZero::SHADER_RGEN, 7, 1>,
+    VkZero::ResourceBinding<VkZero::SwapImage, VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, VkZero::SHADER_RGEN, 8, 1>
   >;
-  using Pipeline = RaytracingPipeline<RaytracingShaderGroup, RaytracingResourceSet>;
-  using RaytracingPushConstantData = PushConstantData<RaytracingPushConstant>;
-  using RaytracingRenderPass_t = RaytracingRenderPass<RaytracingRenderPassPipeline<RaytracingPushConstantData, Pipeline>>;
+  using Pipeline = VkZero::RaytracingPipeline<RaytracingShaderGroup, RaytracingResourceSet>;
+  using RaytracingPushConstantData = VkZero::PushConstantData<RaytracingPushConstant>;
+  using RaytracingRenderPass_t = VkZero::RaytracingRenderPass<VkZero::RaytracingRenderPassPipeline<RaytracingPushConstantData, Pipeline>>;
 
-  SwapImage raytracingStorageImage;
-  SwapImage raytracingPositionStorageImage;
-  SwapImage raytracingLightStorageImageX;
-  SwapImage raytracingLightStorageImageY;
-  SwapImage raytracingLightStorageImageZ;
-  SwapImage raytracingLightStorageImageW;
+  VkZero::SwapImage raytracingStorageImage;
+  VkZero::SwapImage raytracingPositionStorageImage;
+  VkZero::SwapImage raytracingLightStorageImageX;
+  VkZero::SwapImage raytracingLightStorageImageY;
+  VkZero::SwapImage raytracingLightStorageImageZ;
+  VkZero::SwapImage raytracingLightStorageImageW;
 
   RaytracingPushConstants raytracingPushConstants;
   RaytracingShaderGroup group;
