--- include/VkZero/DescriptorPool.hpp
@@ -1,24 +1,0 @@
-#pragma once
-
-#include <vulkan/vulkan_core.h>
-
-class DescriptorPool
-{
-public:
-    DescriptorPool(const DescriptorPool&) = delete;
-    DescriptorPool& operator=(const DescriptorPool&) = delete;
-
-    static DescriptorPool& instance() {
-        static DescriptorPool instance;
-        return instance;
-    }
-
-    VkDescriptorPool get()
-    {
-        return pool;
-    }
-
-private:
-    VkDescriptorPool pool;
-    DescriptorPool();
-};
