--- src/App.cpp
@@ -1,5 +1,0 @@
-#include "App.hpp"
-
-App::App() {}
-App::~App() {}
-void App::run() {}
